srv can unpack data
